src/requirements.txt
uvicorn
torch
torchvision
numpy
opencv-python
pillow
fastapi
deep_utils
jinja2
python-dotenv